+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Improved handling of \n \t and other escaped characters in gdb log window.
 * Show breakpoint info as a tooltip if the breakpoint icon is clicked with
   LMB and held down.
+* Add Cuda information. Very experimental.
 
 ## [2.4] - 2024-03-18
 * Changed main icon to a more license friendly one.